Crate petitset

About

Array-backed ordered set and map data structures in Rust, optimized for stack-allocated storage of a tiny number of elements with a fixed cap. All you need is Eq!

This crate is:

entirely safe

fully documented

no_std compatible

zero dependencies (unless you want thiserror or serde integration)

This crate is designed to be used in performance-sensitive contexts with a small number of elements, where iteration is more common than look-ups and you don’t mind a fixed size. One particularly useful quirk is that elements are not recompacted upon removal: this can be very useful when replacing elements in a set or using the indexes that the elements are stored at in a semantic fashion. Iteration order is guaranteed to be stable, on a first-in-first-out basis.

If this isn’t what you’re after, check out one of these alternatives!

smolset: automatically converts to a HashSet when the number of elements is too large. Unordered.

array_map: Blazing fast, fixed size. All possible keys must be known statically.

sparseset: Heap-allocated, great for sparse data and frequent iteration. Stable order!

HashSet: Heap-allocated, unordered, requires Hash, unbounded size.

BTreeSet: Heap-allocated, ordered, requires Ord, unbounded size.

IndexMap: Heap-allocated, requires Hash, unbounded size.

This crate has a reasonable collection of convenience methods for working with both sets and maps, aiming for rough equivalence with HashMap and HashSet. If you’d like more, please submit an issue or PR!

Modules

set_algebra

Algebraic manipulations of PetitSets

Structs

CapacityError

An error returned when attempting to insert into a full PetitSet or PetitMap.

PetitMap

A map-like data structure with a fixed maximum size

PetitSet

A set-like data structure with a fixed maximum size

Enums

SuccesfulMapInsertion

The Ok result of a successful PetitMap insertion operation

SuccesfulSetInsertion

The Ok result of a successful PetitSet insertion operation